refactor: 이동 규칙에 따른 Piece 추상화 계층 추가
