+++ src/main/java/janggiGame/arrangement/OuterElephantStrategy.java
@@ -2,14 +2,14 @@ package janggiGame.arrangement;
 
 import janggiGame.board.Board;
 import janggiGame.board.Dot;
-import janggiGame.piece.Advisor;
-import janggiGame.piece.Cannon;
-import janggiGame.piece.Chariot;
+import janggiGame.piece.oneMovePiece.Advisor;
+import janggiGame.piece.straightMovePiece.Cannon;
+import janggiGame.piece.straightMovePiece.Chariot;
 import janggiGame.piece.Dynasty;
-import janggiGame.piece.Elephant;
-import janggiGame.piece.Horse;
-import janggiGame.piece.King;
-import janggiGame.piece.Pawn;
+import janggiGame.piece.curveMovePiece.Elephant;
+import janggiGame.piece.curveMovePiece.Horse;
+import janggiGame.piece.oneMovePiece.King;
+import janggiGame.piece.oneMovePiece.Pawn;
 import janggiGame.piece.Piece;
 import java.util.HashMap;
 import java.util.Map;
