+++ src/main/java/janggiGame/piece/oneMovePiece/Advisor.java
@@ -1,0 +1,16 @@
+package janggiGame.piece.oneMovePiece;
+
+import janggiGame.piece.Dynasty;
+
+public class Advisor extends OneMovePiece {
+    public static final String NAME = "사";
+
+    public Advisor(Dynasty dynasty) {
+        super(dynasty);
+    }
+
+    @Override
+    public String getName() {
+        return NAME;
+    }
+}
