+++ src/test/java/janggiGame/piece/ElephantTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 
 import janggiGame.board.Board;
 import janggiGame.board.Dot;
+import janggiGame.piece.curveMovePiece.Elephant;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
