+++ src/test/java/janggiGame/piece/KingTest.java
@@ -2,6 +2,8 @@ package janggiGame.piece;
 
 import janggiGame.board.Board;
 import janggiGame.board.Dot;
+import janggiGame.piece.oneMovePiece.Advisor;
+import janggiGame.piece.straightMovePiece.Chariot;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
